fix no register duplicate user
